+++ docs/CHANGE_LOG.md
@@ -2,6 +2,12 @@
 
 ## Version 0.7.0:
 
+#### **Public Functions:**
+- Added "addAttributes()" support for "html()", which states if the attributes should be written for each element when "write()" is called.
+- Added "addCssProperties()" support for "html()", which states if the CSS properties should be written for each element when "write()" is called.
+- Added "addText()" support for "html()", which states if the text should be written for each element when "write()" is called.
+- Added "addChildren()" support for "html()", which states if the children should be written for each element when "write()" is called.
+
 #### **Fixes & Improvements:**
 - Internal refactoring to make areas a bit easier to read.